+++ MonoRec/Repositories/VisitRepository.cs
@@ -3,6 +3,7 @@ using MonoRec.Data;
 using MonoRec.Models;
 using System.Security.Claims;
 using System.Runtime.Intrinsics.Arm;
+using System.IO;
 
 namespace MonoRec.Repositories
 {
@@ -21,53 +22,72 @@ namespace MonoRec.Repositories
             return visits;
         }
 
-        public Visit GetVisit(int VisitId)
+        public Visit? GetVisit(int VisitId)
         {
-            var visit = _db.Visits.Where(visit => visit.VisitId == VisitId).First();
+            var visit = _db.Visits.FirstOrDefault(visit => visit.VisitId == VisitId);
+
+            if (visit == null) return null;
+
             return visit;
         }
 
-        public Visit CreateNewVisit(int PatientId, int DoctorId)
+        public Visit CreateNewVisit(int patId, int docId)
         {
-            var newVisit = new Visit(PatientId, DoctorId);
+            var newVisit = new Visit(patId, docId);
             _db.Visits.Add(newVisit);
             _db.SaveChanges();
             return newVisit;
         }
 
-        public IEnumerable<Visit> GetAllVisitsByPatient(int PatientId)
+        public IEnumerable<Visit>? GetAllVisitsByPatient(int patId)
         {
+            var patient = _db.Patients.FirstOrDefault(patient => patient.PatientId == patId);
+
+            if (patient == null) return null;
+
             var visits = _db.Visits.ToList();
             var selectResult = from visit in visits
-                               where visit.PatientId == PatientId
+                               where visit.PatientId == patId
                                select visit;
 
             return selectResult;
         }
 
-        public IEnumerable<Visit> GetAllVisitsByDoctor(int DoctorId)
+        public IEnumerable<Visit>? GetAllVisitsByDoctor(int docId)
         {
+            var doctor = _db.Doctors.FirstOrDefault(doctor => doctor.DoctorId == docId);
+
+            if (doctor == null) return null;
+
             var visits = _db.Visits.ToList();
             var selectResult = from visit in visits
-                               where visit.DoctorId == DoctorId
+                               where visit.DoctorId == docId
                                select visit;
 
             return selectResult;
         }
 
-        public IEnumerable<Visit> GetAllVisitsByDoctorPatient(int DoctorId, int PatientId)
+        public IEnumerable<Visit>? GetAllVisitsByDoctorPatient(int docId, int patId)
         {
+            var doctor = _db.Doctors.FirstOrDefault(doctor => doctor.DoctorId == docId);
+            var patient = _db.Patients.FirstOrDefault(patient => patient.PatientId == patId);
+
+            if (doctor == null || patient == null) return null;
+
             var visits = _db.Visits.ToList();
             var selectResult = from visit in visits
-                               where visit.DoctorId == DoctorId && visit.PatientId == PatientId
+                               where visit.DoctorId == docId && visit.PatientId == patId
                                select visit;
 
             return selectResult;
         }
 
-        public Visit DeleteVisit(int VisitId)
+        public Visit? DeleteVisit(int VisitId)
         {
-            var visitToDelete = _db.Visits.Where(visit => visit.VisitId == VisitId).First();
+	        var visitToDelete = _db.Visits.FirstOrDefault(visit => visit.VisitId == VisitId);
+
+            if (visitToDelete == null) return null;
+
             _db.Visits.Remove(visitToDelete);
             _db.SaveChanges();
             return visitToDelete;
